add feature login
added google user authentication and data base access. moved two one file to avoid dependency issues.
