+++ package.json
@@ -4,7 +4,7 @@
   },
   "name": "day4words",
   "description": "A small firebase web app that displays daily vocabulary. Displays the 1000 most common Italian words 4 per day. The counter can manually be changed to the corresponding day. The english translation can be seen by clicking on the word card, as well as the italian pronounciation by tapping the speech button. This website is currently hosted under day4words.web.app.",
-  "version": "1.0.2",
+  "version": "1.1.0",
   "main": "index.js",
   "scripts": {
     "test": "firebase serve"
