+++ public/index.html
@@ -11,7 +11,6 @@
 
   <link rel="stylesheet" href="style.css">
 
-  <script type="module" src="words.js"></script>
   <script type="module" src="index.js"></script>
   <script type="module" src="speech.js"></script>
 
@@ -49,7 +48,7 @@
   <input type="text" class="counter" value="0" id="counter"></input>
   <button class="button" id="minus">-</button>
   <button class="button" id="plus">+</button>
-  <p style="float:right"> v1.0.2</p>
+  <p style="float:right"> v1.1.0</p>
